About Us Seeder Add & Controller Update

+++ app/Http/Controllers/Admin/AboutUsController.php
@@ -55,7 +55,11 @@ class AboutUsController extends Controller
                 ]
             );
 
-            if (file_exists($oldImage)) {
+            $defaultImages = [
+                'frontend/images/about/about-page.jpg',
+            ];
+
+            if ($oldImage && !in_array($oldImage, $defaultImages) && file_exists($oldImage)) {
                 unlink($oldImage);
             }
